switch to m13-es-transformer
while

  @zerollup/ts-transform-paths
  @zoltu/typescript-transformer-append-js-extension

are working again with recent versions of typescript, they cause the watch mode to exit
with an exception. so i copied

  m13-es-transformer

from toad.js, and while it floods the console with

  (node:6515) Warning: Accessing non-existent property 'tracing' of module exports inside circular dependency

which i haven't been able to track down, it seems to work for now.

this also removes the "type": "module" from the package.json.